:recycle: Move filterContainer to components/shared
