+++ src/components/shared/filterDropdownButton/filters.css
@@ -1,5 +1,5 @@
-@import '../../../../../app/variables.css';
-@import '../../../../../app/mixins.css';
+@import '../../../app/variables.css';
+@import '../../../app/mixins.css';
 
 .fieldGroup {
   display: flex;
